Optimize CPU-bound Task Execution by Replacing Thread Lock with Producer-Consumer Pattern (#957)
## Issue
Seeing large slowdowns when using `evaluate()` compared to naive
`ThreadPoolExecutor`. The specific slowdown happens on a cpu-bound call
to BeautifulSoup, which balloons from 0.1s -> 5 seconds (and sometimes
up to 60+ seconds based on concurrency).

## Profiling Investigation
First, I noticed that the majority of the time in bs4 init was actually
empty space, which made me think that: (1) there was GIL contention, or
(2) there was some lock being placed on the thread.
![CleanShot 2024-08-28 at 13 09
14](https://github.com/user-attachments/assets/85f9b9a6-a401-405c-a428-d573e31c14b5)

I took a look at the `ExperimentResults` implementation, and noted that
in the infinite while loop, there was an `is_active` call inside of the
thread lock, and since `is_active` has nontrivial overhead, and this
loop was running even if there were no results, we were incurring this
overhead over and over.
![CleanShot 2024-08-28 at 13 10
33](https://github.com/user-attachments/assets/3525a57d-10d2-4ab6-9801-7b852543b4c2)

To confirm, I noticed that in the main thread profile, about half the
time is spent in `is_alive` in aggregate.
![CleanShot 2024-08-28 at 13 13
04](https://github.com/user-attachments/assets/45b84f28-e3d2-4e51-b565-dc168220af7b)

## Fix
To improve the implementation without blocking CPU-bound operations, we
can use a producer-consumer pattern with a queue. This will allow the
main thread to continue processing while the data is being fetched and
processed in the background.

---------

Co-authored-by: William Fu-Hinthorn <13333726+hinthornw@users.noreply.github.com>