Remove links in vendored code that point outside the sdist source code. (#1256)
We vendor `pyo3` and its related libraries, but we depend on users'
build tools to build wheels when a suitable wheel is not found on PyPI.
The source distribution only includes the actual source code of `pyo3`
et al, but their source code indirectly depends on non-source artifacts,
such as files in their `guide/` or `examples/` directories. Those
directories are not part of the source distribution, so locally building
the wheel fails.

This PR removes those links, since they are only used for doc comments
and aren't relevant to the actual operation of our library.