patch[python]: Don't set metadata values if the value is None (#1641)
This change fixes an issue in the LangSmith OTEL integration where
metadata attributes with a value of None were being set on spans,
causing errors like:

Invalid type NoneType for attribute 'langsmith.metadata.ls_temperature'
value.

Changes:

Added a check to ensure that metadata values are only set if they are
not None.

Bumped the package version from 0.3.24 to 0.3.25.

Resolves: https://github.com/langchain-ai/langsmith-sdk/issues/1640