patch[python, js]: Update runs with project name (#1659)
Currently, patch requests don't carry information about the project they
are associated with. This causes upstream issues on ingestion when the
patch is sent separately and we don't know the trace tier of the
project, and thus, assign it the default tracing project tier. The
result of this is that we upgrade the s3 urls post ingestion API:
resulting in higher storage costs and an increase in class A operations.